Make DiscFilterRequest non-abstract and deprecate JdiscFilterRequest
There is no need for having DiscFilterRequest as an abstract class after removing servlet support.
