+++ container-core/abi-spec.json
@@ -2164,15 +2164,14 @@
     "superClass": "java.lang.Object",
     "interfaces": [],
     "attributes": [
-      "public",
-      "abstract"
+      "public"
     ],
     "methods": [
       "public void <init>(com.yahoo.jdisc.http.HttpRequest)",
-      "public abstract java.lang.String getMethod()",
+      "public java.lang.String getMethod()",
       "public com.yahoo.jdisc.http.HttpRequest$Version getVersion()",
       "public java.net.URI getUri()",
-      "public abstract void setUri(java.net.URI)",
+      "public void setUri(java.net.URI)",
       "public com.yahoo.jdisc.http.HttpRequest getParentRequest()",
       "public java.lang.String getRemoteAddr()",
       "public void setRemoteAddr(java.lang.String)",
@@ -2182,8 +2181,8 @@
       "public void setAttribute(java.lang.String, java.lang.Object)",
       "public boolean containsAttribute(java.lang.String)",
       "public void removeAttribute(java.lang.String)",
-      "public abstract java.lang.String getParameter(java.lang.String)",
-      "public abstract java.util.Enumeration getParameterNames()",
+      "public java.lang.String getParameter(java.lang.String)",
+      "public java.util.Enumeration getParameterNames()",
       "public java.util.List getParameterNamesAsList()",
       "public java.util.Enumeration getParameterValues(java.lang.String)",
       "public java.util.List getParameterValuesAsList(java.lang.String)",
@@ -2194,16 +2193,16 @@
       "public java.util.Map getUntreatedParams()",
       "public com.yahoo.jdisc.HeaderFields getUntreatedHeaders()",
       "public java.util.List getUntreatedCookies()",
-      "public abstract void addHeader(java.lang.String, java.lang.String)",
+      "public void addHeader(java.lang.String, java.lang.String)",
       "public long getDateHeader(java.lang.String)",
-      "public abstract java.lang.String getHeader(java.lang.String)",
-      "public abstract java.util.Enumeration getHeaderNames()",
-      "public abstract java.util.List getHeaderNamesAsList()",
-      "public abstract java.util.Enumeration getHeaders(java.lang.String)",
-      "public abstract java.util.List getHeadersAsList(java.lang.String)",
-      "public abstract void removeHeaders(java.lang.String)",
-      "public abstract void setHeaders(java.lang.String, java.lang.String)",
-      "public abstract void setHeaders(java.lang.String, java.util.List)",
+      "public java.lang.String getHeader(java.lang.String)",
+      "public java.util.Enumeration getHeaderNames()",
+      "public java.util.List getHeaderNamesAsList()",
+      "public java.util.Enumeration getHeaders(java.lang.String)",
+      "public java.util.List getHeadersAsList(java.lang.String)",
+      "public void removeHeaders(java.lang.String)",
+      "public void setHeaders(java.lang.String, java.lang.String)",
+      "public void setHeaders(java.lang.String, java.util.List)",
       "public int getIntHeader(java.lang.String)",
       "public com.yahoo.container.jdisc.RequestView asRequestView()",
       "public java.util.List getCookies()",
@@ -2218,20 +2217,20 @@
       "public void setScheme(java.lang.String, boolean)",
       "public java.lang.String getServerName()",
       "public int getServerPort()",
-      "public abstract java.security.Principal getUserPrincipal()",
+      "public java.security.Principal getUserPrincipal()",
       "public boolean isSecure()",
       "public boolean isUserInRole(java.lang.String)",
       "public void setOverrideIsUserInRole(boolean)",
       "public void setRemoteHost(java.lang.String)",
       "public void setRemoteUser(java.lang.String)",
-      "public abstract void setUserPrincipal(java.security.Principal)",
-      "public abstract java.util.List getClientCertificateChain()",
+      "public void setUserPrincipal(java.security.Principal)",
+      "public java.util.List getClientCertificateChain()",
       "public void setUserRoles(java.lang.String[])",
       "public java.lang.String getContentType()",
       "public java.lang.String getCharacterEncoding()",
       "public void setCharacterEncoding(java.lang.String)",
       "public void addCookie(com.yahoo.jdisc.http.filter.JDiscCookieWrapper)",
-      "public abstract void clearCookies()",
+      "public void clearCookies()",
       "public com.yahoo.jdisc.http.filter.JDiscCookieWrapper[] getWrappedCookies()",
       "public static boolean isMultipart(com.yahoo.jdisc.http.filter.DiscFilterRequest)"
     ],
@@ -2323,24 +2322,7 @@
     ],
     "methods": [
       "public void <init>(com.yahoo.jdisc.http.HttpRequest)",
-      "public com.yahoo.jdisc.http.HttpRequest getParentRequest()",
-      "public void setUri(java.net.URI)",
-      "public java.lang.String getMethod()",
-      "public java.lang.String getParameter(java.lang.String)",
-      "public java.util.Enumeration getParameterNames()",
-      "public void addHeader(java.lang.String, java.lang.String)",
-      "public java.lang.String getHeader(java.lang.String)",
-      "public java.util.Enumeration getHeaderNames()",
-      "public java.util.List getHeaderNamesAsList()",
-      "public java.util.Enumeration getHeaders(java.lang.String)",
-      "public java.util.List getHeadersAsList(java.lang.String)",
-      "public void removeHeaders(java.lang.String)",
-      "public void setHeaders(java.lang.String, java.lang.String)",
-      "public void setHeaders(java.lang.String, java.util.List)",
-      "public java.security.Principal getUserPrincipal()",
-      "public void setUserPrincipal(java.security.Principal)",
-      "public java.util.List getClientCertificateChain()",
-      "public void clearCookies()"
+      "public com.yahoo.jdisc.http.HttpRequest getParentRequest()"
     ],
     "fields": []
   },
